\name{ardBuck}
\alias{ardBuck}
\title{Arden Buck Equation from Buck Research Manual (1996)
http://cires.colorado.edu/~voemel/vp.html}
\usage{
ardBuck(Tcelsius)
}
\arguments{
  \item{Tcelsius}{the temperature in degrees C}
}
\description{
Arden Buck Equation from Buck Research Manual (1996)
http://cires.colorado.edu/~voemel/vp.html
}

