% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biocro_fba.r
\name{biocro_fba}
\alias{biocro_fba}
\title{biocro_fba}
\usage{
biocro_fba(
  initial_values = list(),
  parameters = list(),
  drivers,
  direct_module_names = list(),
  differential_module_names = list(),
  ode_solver = default_ode_solver,
  verbose = FALSE
)
}
\value{
a dataframe
}
\description{
A function to run biocro while taking metabolism into account
}
\examples{
biocro_fba()
}
