\name{BioGro}
\alias{BioGro}
\title{BioGro}
\usage{
  BioGro(WetDat, day1 = NULL, dayn = NULL, timestep = 1,
    lat = 40, iRhizome = 7, irtl = 1e-04,
    canopyControl = list(), seneControl = list(),
    photoControl = list(), phenoControl = list(),
    soilControl = list(), nitroControl = list(),
    sugarphenoControl = list(), centuryControl = list())
}
\arguments{
  \item{WetDat}{}

  \item{day1}{}

  \item{dayn}{}

  \item{timestep}{}

  \item{lat}{}

  \item{iRhizome}{}

  \item{irtl}{}

  \item{canopyControl}{}

  \item{seneControl}{}

  \item{photoControl}{}

  \item{phenoControl}{}

  \item{soilControl}{}

  \item{nitroControl}{}

  \item{sugarphenoControl}{}

  \item{centuryControl}{}
}
\value{
  result of BioGro function
}
\description{
  .. content for \description{} (no empty lines) ..
}
\details{
  .. content for \details{} ..
}

