\name{CenturyC}
\alias{CenturyC}
\title{C version of the Century function}
\usage{
CenturyC(LeafL, StemL, RootL, RhizL, smoist, stemp, precip, leachWater,
  centuryControl = list(), soilType = 0)
}
\arguments{
  \item{LeafL}{Leaf litter.}

  \item{StemL}{Stem litter.}

  \item{RootL}{Root litter.}

  \item{RhizL}{Rhizome litter.}

  \item{smoist}{Soil moisture.}

  \item{stemp}{Soil temperature.}

  \item{precip}{Precipitation.}

  \item{leachWater}{Leached water.}

  \item{centuryControl}{See \code{\link{centuryParms}}.}

  \item{soilType}{See \code{\link{showSoilType}}.}
}
\description{
C version of the Century function
}

