% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CheckLeapYear.R
\name{CheckLeapYear}
\alias{CheckLeapYear}
\title{This fuction checks if year is a leap year. If yes, then returns 1 or 0.}
\usage{
CheckLeapYear(year)
}
\description{
This fuction checks if year is a leap year. If yes, then returns 1 or 0.
}
