\name{cmi0506}
\docType{data}
\alias{cmi0506}
\title{Weather data}
\description{
selected weather data corresponding to the Champaign weather
station (IL, U.S.). It has two years: 2005 and 2006.
Dimensions: 730 by 11. The columns correspond to the input necessary for the
\code{\link{weach}} function.
}

\usage{cmi0506}
\format{data frame of dimensions 730 by 11.}
\source{\url{http://www.sws.uiuc.edu/warm/}}
\keyword{datasets}