% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biocro-package.R
\docType{data}
\name{cmiWet}
\alias{cmiWet}
\title{Weather data}
\format{data frame of dimensions 384 by 9.}
\source{
simulated
}
\description{
Layer data for evapo/transpiration. Simulated data to show the result of
the EvapoTrans function.
}
\details{
lfClass: leaf class, 'sun' or 'shade'.

layer: layer in the canopy, 1 to 8.

hour: hour of the day, (0--23).

Rad: direct light.

Itot: total radiation.

Temp: air temperature, (Celsius).

RH: relative humidity, (0--1).

WindSpeed: wind speed, (m \eqn{s^{-1}}).

LAI: leaf area index.
}
\keyword{datasets}
