% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.configs.BIOCRO.R
\name{convert.samples.BIOCRO}
\alias{convert.samples.BIOCRO}
\title{convert parameters from PEcAn database default units to biocro defaults}
\usage{
convert.samples.BIOCRO(trait.samples, biocro_version = 1)
}
\arguments{
\item{trait.samples}{a matrix or dataframe of samples from the trait distribution}

\item{biocro_version}{numeric, but currently only checks whether version is less than 1.0}
}
\value{
dataframe with values transformed
}
\description{
Performs model specific unit conversions on a list of trait values,
such as those provided to write.config
}
\author{
David LeBauer
}
