% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/compile_doxygen.R
\name{document}
\alias{document}
\title{devtools document function when using doxygen}
\usage{
document()
}
\value{
The value returned by devtools::document()
}
\description{
Overwrites devtools::document() to include the treatment of
   doxygen documentation in src/
}

