% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biocro-package.R
\docType{data}
\name{doy124}
\alias{doy124}
\title{Weather data}
\format{Data frame of dimensions 24 by 8.}
\source{
Simulated.
}
\description{
Example for a given day of the year to illustrate the \code{\link{CanA}}
function.
}
\details{
LAI: Leaf area index.

year: Year.

doy: 124 in this case.

hour: Hour of the day (0--23).

solarR: Direct solar radiation.

DailyTemp.C: Hourly air temperature (Celsius).

RH: Relative humidity (0--1).

WindSpeed: 4.1 m s\eqn{^{-1}}{-1} average daily value in this case.
}
\keyword{datasets}
