\name{fnpsvp}
\alias{fnpsvp}
\title{The Goff Gratch equation from Smithsonian Tables, 1984.
http://cires.colorado.edu/~voemel/vp.html}
\usage{
fnpsvp(Tkelvin)
}
\arguments{
  \item{Tkelvin}{the abolute temperature}
}
\description{
The Goff Gratch equation from Smithsonian Tables, 1984.
http://cires.colorado.edu/~voemel/vp.html
}

