% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/get.model.output.BIOCRO.R
\name{get.model.output.BIOCRO}
\alias{get.model.output.BIOCRO}
\title{Retrieve model output from local server}
\usage{
get.model.output.BIOCRO(settings)
}
\arguments{
\item{settings}{list generated from \code{\link{read.settings}} function applied to settings file}
}
\description{
Function to retrieve model output from local server
}
\author{
Mike Dietze, David LeBauer
}

