% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_biocro_defaults.R
\name{get_biocro_defaults}
\alias{get_biocro_defaults}
\title{Look for defaults provided as datasets in the BioCro model package}
\usage{
get_biocro_defaults(genus)
}
\arguments{
\item{genus}{Name of the genus (or really any string BioCro uses as a *_parameters prefix)}
}
\value{
a list in the format expected by `BioCro::Gro()`,
  containing four lists named `type`, `initial_values`, `parameters`, and `modules`,
  or NULL if genus not found
}
\description{
When available, default parameters come in sets of three:
*_initial_values, *_parameters, *_modules
}
