\name{LayET}
\docType{data}
\alias{LayET}
\title{Weather data}
\description{
Layer data for evapo/transpiration. Simulated data to show the
result of the EvapoTrans function.

lfClass: leaf class, "sun" or "shade".

layer: layer in the canopy, 1 to 8.

hour: hour of the day, (0--23).

Rad: direct light.

Itot: total radiation.

Temp: air temperature, (Celsius).

RH: relative humidity, (0--1).

WindSpeed: wind speed, (m \eqn{s^{-1}}).

LAI: leaf area index.
}

\usage{LayET}
\format{data frame of dimensions 384 by 9.}
\source{simulated}
\keyword{datasets}