% Generated by roxygen2 (4.0.1): do not edit by hand
\name{MaizeGro}
\alias{MaizeGro}
\title{Simulation of Maize, Growth, LAI, Photosynthesis and phenology}
\usage{
MaizeGro(WetDat, plant.day = NULL, emerge.day = NULL, harvest.day = NULL,
  plant.density = 7, timestep = 1, lat = 40, canopyControl = list(),
  MaizeSeneControl = list(), photoControl = list(),
  MaizePhenoControl = list(), MaizeCAllocControl = list(),
  laiControl = list(), soilControl = list(), MaizeNitroControl = list(),
  centuryControl = list())
}
\arguments{
\item{WetDat}{weather data as produced by the \code{\link{weach}} function.}

\item{plant.day}{Planting date (format 0-365)}

\item{emerge.day}{Emergence date (format 0-365)}

\item{harvest.day}{Harvest date (format 0-365)}

\item{plant.density}{Planting density (plants per meter squared, default =
7)}

\item{timestep}{Simulation timestep, the default of 1 requires houlry
weather data. A value of 3 would require weather data every 3 hours.  This
number should be a divisor of 24.}

\item{lat}{latitude, default 40.}

\item{canopyControl}{List that controls aspects of the canopy simulation. It
should be supplied through the \code{canopyParms} function.

\code{Sp} (specific leaf area) here the units are ha \eqn{Mg^{-1}}.  If you
have data in \eqn{m^2} of leaf per kg of dry matter (e.g. 15) then divide by
10 before inputting this coefficient.

\code{SpD} decrease of specific leaf area. Empirical parameter. Default 0.
example value (1.7e-3).

\code{nlayers} (number of layers of the canopy) Maximum 50. To increase the
number of layers (more than 50) the \code{C} source code needs to be changed
slightly.

\code{kd} (extinction coefficient for diffuse light) between 0 and 1.

\code{mResp} (maintenance respiration) a vector of length 2 with the first
component for leaf and stem and the second component for rhizome and root.}

\item{MaizeSeneControl}{List that controls aspects of senescence simulation.
It should be supplied through the \code{MaizeSeneParms} function.

\code{senLeaf} Thermal time at which leaf senescence will start.

\code{senStem} Thermal time at which stem senescence will start.

\code{senRoot} Thermal time at which root senescence will start.}

\item{photoControl}{List that controls aspects of photosynthesis simulation.
It should be supplied through the \code{MaizePhotoParms} function.

\code{vmax} Vmax passed to the \code{\link{c4photo}} function.

\code{alpha} alpha parameter passed to the \code{\link{c4photo}} function.

\code{kparm} kparm parameter passed to the \code{\link{c4photo}} function.

\code{theta} theta parameter passed to the \code{\link{c4photo}} function.

\code{beta} beta parameter passed to the \code{\link{c4photo}} function.

\code{Rd} Rd parameter passed to the \code{\link{c4photo}} function.

\code{UPPERTEMP} UPPERTEMP parameter passed to the \code{\link{c4photo}}
function.

\code{LOWERTEMP} LOWERTEMP parameter passed to the \code{\link{c4photo}}
function.

\code{Catm} Catm parameter passed to the \code{\link{c4photo}} function.

\code{b0} b0 parameter passed to the \code{\link{c4photo}} function.

\code{b1} b1 parameter passed to the \code{\link{c4photo}} function.}

\item{MaizePhenoControl}{argument used to pass parameters related to
phenology characteristics }

\item{soilControl}{ see BioGro function }

\item{nitroControl}{ see BioGro function }

\item{centuryControl}{ see BioGro function }
}
\value{
It currently returns a list with the following components

\item{DayofYear}{Day of the year (0-365)}

\item{Hour}{Hour of the day (0-23)}

\item{TTTc}{Accumulated thermal time}

\item{PhenoStage}{Phenological stage of the crop}

\item{CanopyAssim}{Hourly canopy assimilation, (Mg \eqn{ha^-1} ground
\eqn{hr^-1}).}

\item{CanopyTrans}{Hourly canopy transpiration, (Mg \eqn{ha^-1} ground
\eqn{hr^-1}).}

\item{LAI}{Leaf Area Index}
}
\description{
It takes weather data as input (hourly timesteps) and several parameters and
it produces phenology, photosynthesis, LAI, etc.
}
\details{
The phenology follows the 'Corn Growth and Development' Iowa State
Publication. 
}
\examples{
data(weather05)
res <- MaizeGro(weather05, plant.day = 110, emerge.day = 120, harvest.day=300,
                  MaizePhenoControl = MaizePhenoParms(R6 = 2000))
}
\author{
Fernando E Miguez
}

\seealso{
\code{\link{BioGro}} 
}
\keyword{models}

