% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MaizeGro.R
\name{MaizePhenoParms}
\alias{MaizePhenoParms}
\title{Lists values for phenology parameters}
\usage{
MaizePhenoParms(base.temp = 10, max.leaves = 20, plant.emerg = 100,
  phyllochron1 = 46.7, phyllochron2 = 31.1, R1 = 747, R2 = 858,
  R3 = 969, R4 = 1080, R5 = 1136, R6 = 1452)
}
\arguments{
\item{base.temp}{inital temperatre}

\item{max.leaves}{maximum leaves}

\item{plant.emerg}{plant emergence}

\item{phyllochron1}{needs description}

\item{phyllochron2}{needs description}

\item{R1}{needs description}

\item{R2}{needs description}

\item{R3}{needs description}

\item{R4}{needs description}

\item{R5}{needs description}

\item{R6}{needs description}
}
\description{
Lists values for phenology parameters
}

