% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/met2model.BIOCRO.R
\name{met2model.BIOCRO}
\alias{met2model.BIOCRO}
\title{Write BioCro met files}
\usage{
met2model.BIOCRO(in.path, in.prefix, outfolder, overwrite = FALSE, lat,
  lon, start_date, end_date, ...)
}
\arguments{
\item{in.path}{path on disk where CF file lives}

\item{in.prefix}{prefix for each file}

\item{outfolder}{location where model specific output is written}

\item{overwrite}{logical: Write over any existing file of the same name? If FALSE, leaves the existing file untouched and skips to the next year.}

\item{lat, lon}{Site latitude and longitude}

\item{start_date, end_date}{Date range to convert. Each year will be written to a separate file}

\item{...}{other arguments passed from PEcAn, currently ignored}
}
\value{
a dataframe of information about the written file
}
\description{
Converts a met CF file to a model specific met file. The input
files are called <in.path>/<in.prefix>.YYYY.cf
}
\author{
Rob Kooper, David LeBauer
}
