% Generated by roxygen2 (4.0.2): do not edit by hand
\name{met2model.MODEL}
\alias{met2model.BIOCRO}
\alias{met2model.MODEL}
\title{Write MODEL met files}
\usage{
met2model.BIOCRO(in.path, in.prefix, outfolder, overwrite = FALSE)
}
\arguments{
\item{in.path}{path on disk where CF file lives}

\item{in.prefix}{prefix for each file}

\item{outfolder}{location where model specific output is written.}
}
\value{
OK if everything was succesful.
}
\description{
Converts a met CF file to a model specific met file. The input
files are calld <in.path>/<in.prefix>.YYYY.cf
}
\author{
Rob Kooper
}

