\docType{data}
\name{obsBea}
\alias{obsBea}
\title{Miscanthus assimilation field data}
\format{data frame of dimensions 27 by 4.}
\source{
  C. V. Beale, D. A. Bint, S. P. Long, Leaf photosynthesis
  in the C4-grass miscanthus x giganteus, growing in the
  cool temperate climate of southern England, \emph{J. Exp.
  Bot.} 47 (2) (1996) 267--273.
}
\description{
  assimilation as measured in Beale, Bint and Long (1996)
  in Miscanthus.  The first column is the observed net
  assimilation rate (\eqn{\mu} mol \eqn{m^{-2}}
  \eqn{s^{-1}}).  The second column is the observed quantum
  flux (\eqn{\mu} mol \eqn{m^{-2}} \eqn{s^{-1}}). The third
  column is the temperature (Celsius).Relative humidity was
  not reported and thus was assumed to be 0.7.
}
\keyword{datasets}

