\name{phenoParms}
\alias{phenoParms}
\title{phenoparms}
\usage{
  phenoParms(tp1 = 562, tp2 = 1312, tp3 = 2063, tp4 = 2676,
    tp5 = 3211, tp6 = 7000, kLeaf1 = 0.33, kStem1 = 0.37,
    kRoot1 = 0.3, kRhizome1 = -8e-04, kLeaf2 = 0.14,
    kStem2 = 0.85, kRoot2 = 0.01, kRhizome2 = -5e-04,
    kLeaf3 = 0.01, kStem3 = 0.63, kRoot3 = 0.01,
    kRhizome3 = 0.35, kLeaf4 = 0.01, kStem4 = 0.63,
    kRoot4 = 0.01, kRhizome4 = 0.35, kLeaf5 = 0.01,
    kStem5 = 0.63, kRoot5 = 0.01, kRhizome5 = 0.35,
    kLeaf6 = 0.01, kStem6 = 0.63, kRoot6 = 0.01,
    kRhizome6 = 0.35, kGrain6 = 0)
}
\arguments{
  \item{tp1}{}

  \item{tp2}{}

  \item{tp3}{}

  \item{tp4}{}

  \item{tp5}{}

  \item{tp6}{}

  \item{kLeaf1}{}

  \item{kStem1}{}

  \item{kRoot1}{}

  \item{kRhizome1}{}

  \item{kLeaf2}{}

  \item{kStem2}{}

  \item{kRoot2}{}

  \item{kRhizome2}{}

  \item{kLeaf3}{}

  \item{kStem3}{}

  \item{kRoot3}{}

  \item{kRhizome3}{}

  \item{kLeaf4}{}

  \item{kStem4}{}

  \item{kRoot4}{}

  \item{kRhizome4}{}

  \item{kLeaf5}{}

  \item{kStem5}{}

  \item{kRoot5}{}

  \item{kRhizome5}{}

  \item{kLeaf6}{}

  \item{kStem6}{}

  \item{kRoot6}{}

  \item{kRhizome6}{}

  \item{kGrain6}{}
}
\value{
  phenological parameters
}
\description{
  .. content for \description{} (no empty lines) ..
}
\details{
  .. content for \details{} ..
}
\author{
  Fernando Miguez
}

