\name{print.OpBioGro}
\alias{print.OpBioGro}
\title{ Print an OpBioGro object}
\description{
Print an object generated by OpBioGro
}
\usage{
\method{print}{OpBioGro}(x, ...)
}
\arguments{
\item{x}{
  Object returned from BioGro optimization
  (Collatz model).
 }
\item{\dots}{
  Optional arguments. 
 }
}
\seealso{  \code{\link{OpBioGro}}}

\keyword{optimize}
