% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Opc3photo.R
\name{print.Opc3photo}
\alias{print.Opc3photo}
\title{Display methods for Opc4photo and OpEC4photo}
\usage{
\method{print}{Opc3photo}(x, digits = 2, ...)
}
\description{
Display methods for Opc4photo and OpEC4photo
}

