% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Opc4photo.R
\name{print.Opc4photo}
\alias{print.Opc4photo}
\title{Display methods for Opc4photo and OpEC4photo}
\usage{
\method{print}{Opc4photo}(x, digits = 2, ...)
}
\description{
Display methods for Opc4photo and OpEC4photo
}
