\name{print.OpEC4photo}
\alias{print.OpEC4photo}
\title{ Print an OpEC4photo object}
\description{
Print an object generated by OpEC4photo
}
\usage{
\method{print}{OpEC4photo}(x, \dots)
}
\arguments{
\item{x}{
  Object returned from C4 photosynthesis optimization
(von Caemmerer model).
 }
\item{\dots}{
  Optional arguments. 
 }
}
\seealso{  \code{\link{OpEC4photo}}}

\keyword{optimize}
