% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.biocro.config.R
\name{read.biocro.config}
\alias{read.biocro.config}
\title{Read BioCro config file}
\usage{
read.biocro.config(config.file = "config.xml")
}
\arguments{
\item{config.file}{Path to XML file}
}
\value{
list of run configuration parameters for PEcAn
}
\description{
Read BioCro config file
}
\author{
David LeBauer
}
