% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/write.configs.BIOCRO.R
\name{remove.config.BIOCRO}
\alias{remove.config.BIOCRO}
\title{Clear out previous biocro config and parameter files.}
\usage{
remove.config.BIOCRO(main.outdir, settings)
}
\arguments{
\item{main.outdir}{Primary PEcAn output directory (will be depreciated)}

\item{settings}{PEcAn settings file}
}
\value{
nothing, removes config files as side effect
}
\description{
Clear out previous config and parameter files.
}
\author{
Shawn Serbin, David LeBauer
}

