\name{RssMaizeGro}
\alias{RssMaizeGro}
\title{Very simple function to compare the distance between simulated
and observed data for the BioGro function
Need to add an argument such as pc.sigmas "plant component sigmas"
If variability of the plant component is known}
\usage{
RssMaizeGro(obs, sim)
}
\description{
Very simple function to compare the distance between
simulated and observed data for the BioGro function Need to
add an argument such as pc.sigmas "plant component sigmas"
If variability of the plant component is known
}

