\docType{data}
\name{simDat2}
\alias{simDat2}
\title{Simulated biomass data.}
\format{A data frame with 5 observations on the following 6 variables.
    \describe{
        \item{list('TT')}{A numeric vector.}
        \item{list('Stem')}{A numeric vector.}
        \item{list('Leaf')}{A numeric vector.}
        \item{list('Root')}{A numeric vector.}
        \item{list('Rhiz')}{A numeric vector.}
        \item{list('LAI')}{A numeric vector.}
    }
}
\source{
}
\description{
}
\details{
}
\examples{

data(simDat2)
## maybe str(simDat2) ; plot(simDat2) ...

}
\references{
}
\keyword{datasets}
