\name{SoilType}
\alias{SoilType}
\title{Soil Type}
\usage{
  SoilType(soiltype)
}
\arguments{
  \item{soiltype}{integer between 0 and 10}
}
\value{
  list with soil name and physical properties
}
\description{
  .. content for \description{} (no empty lines) ..
}
\details{
  .. content for \details{} ..
}
\author{
  Fernando Miguez
}

