\name{start.runs.BIOCRO}
\alias{start.runs.BIOCRO}
\title{Start run of biocro model}
\usage{
  start.runs.BIOCRO(runid)
}
\arguments{
  \item{runid}{the id of the run (folder in runs) to
  execute}
}
\description{
  Start biocro model runs on local
}
\author{
  Rob Kooper, David LeBauer, Deepak Jaiswal
}

