% Generated by roxygen2 (4.0.1): do not edit by hand
\name{summary.Opc4photo}
\alias{summary.Opc4photo}
\title{This function will implement simple calculations
of predicted and residuals for the Opc4photo function}
\usage{
\method{summary}{Opc4photo}(object, ...)
}
\description{
This function will implement simple calculations
of predicted and residuals for the Opc4photo function
}

