\name{TempToDdryA}
\alias{TempToDdryA}
\title{Returns a value for TempToDdryA}
\usage{
TempToDdryA(Temp)
}
\arguments{
  \item{Temp}{Temperature}
}
\description{
Takes a value for Temp as defined by the SoilEvapo function
and returns a value for DdryA which factors into variable
PsychParam which in turn helps define the Evaporation.
}

