\name{TempToLHV}
\alias{TempToLHV}
\title{takes a value for Temp as defined by the SoilEvapo function and returns a value for LHV
which factors into variable PsychParam which in turn helps define the Evaporation.}
\usage{
TempToLHV(Temp)
}
\description{
takes a value for Temp as defined by the SoilEvapo function
and returns a value for LHV which factors into variable
PsychParam which in turn helps define the Evaporation.
}

