\name{TempToSFS}
\alias{TempToSFS}
\title{takes a value for Temp as defined by the SoilEvapo function and returns a value for SlopeFS
which helps define the Evaporation.}
\usage{
TempToSFS(Temp)
}
\description{
takes a value for Temp as defined by the SoilEvapo function
and returns a value for SlopeFS which helps define the
Evaporation.
}

