\name{TempToSWVC}
\alias{TempToSWVC}
\title{Returns a value for TempToSWC}
\usage{
TempToSWVC(Temp)
}
\arguments{
  \item{Temp}{Temperature}
}
\description{
Takes a value for Temp as defined by the SoilEvapo function
and returns a value for SWVC which factors into variable
DeltaPVa which in turn helps define the Evaporation.
}

