\name{weachNEW}
\alias{weachNEW}
\title{weachNEW}
\usage{
  weachNEW(X, lati, ts = 1,
    temp.units = c("Farenheit", "Celsius"),
    rh.units = c("percent", "fraction"),
    ws.units = c("mph", "mps"), pp.units = c("in", "mm"),
    ...)
}
\arguments{
  \item{X}{}

  \item{lati}{}

  \item{ts}{}

  \item{temp.units}{}

  \item{rh.units}{}

  \item{ws.units}{}

  \item{pp.units}{}

  \item{...}{}
}
\value{
  data.frame with NCEP met data
}
\description{
  like weach, but new
}
\author{
  Fernando Miguez, Deepak Jaiswal
}

