\name{weather05}
\docType{data}
\alias{weather05}
\alias{weather04}
\title{Weather data}
\description{
  Weather data as produced by the \code{\link{weach}} function.
  These are for 2004 and 2005.
}
\usage{weather05}
\format{data frame of dimensions 8760 by 7.}
\source{simulated (based on Champaign, Illinois
conditions).}
\keyword{datasets}