\name{wide2long}
\alias{wide2long}
\title{Wide to Long}
\usage{
  wide2long(data.wide, lat, lon, var)
}
\arguments{
  \item{data.wide}{data}

  \item{lat}{latitude for rows}

  \item{lon}{longitude for columns}

  \item{var}{variable being measured}
}
\value{
  data.frame with colnames (lat, lon, var)
}
\description{
  .. content for \description{} (no empty lines) ..
}
\details{
  .. content for \details{} ..
}
\author{
  David LeBauer
}

