\name{wsRcoef}
\alias{wsRcoef}
\title{determines whether the argument wsFun is linear, logistic,
exponential, or something else and calculates a value for wsPhoto
based on that.}
\usage{
wsRcoef(aw, fieldc, wiltp, phi1, phi2, wsFun = c("linear", "logistic", "exp",
  "none"))
}
\description{
determines whether the argument wsFun is linear, logistic,
exponential, or something else and calculates a value for
wsPhoto based on that.
}

