\name{catm_data}

\docType{data}

\alias{catm_data}

\title{Global annual mean atmopspheric CO2 levels}

\description{
  Multiple years of global annual mean atmospheric CO2 levels and their
  uncertainties.
}

\format{
  Data frame with 3 columns and 41 rows.
}

\source{
  Data were obtained from the National Oceanic and Atmospheric Administration's
  Global Monitoring Laboratory (https://gml.noaa.gov/ccgg/trends/data.html) on
  04-13-2021.
}

\keyword{weather}
