% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/met2model.BIOCRO.R
\name{cf2biocro}
\alias{cf2biocro}
\title{Convert CF data.table to BioCro input}
\usage{
cf2biocro(met)
}
\arguments{
\item{met}{data.table with met for a single site; output from \link{\code{load.cfmet}}}

\item{in.prefix}{prefix for each file}

\item{outfolder}{location where model specific output is written.}

\item{...}{can pass lat, lon, start.date and end.date}
}
\value{
OK if everything was succesful.
}
\description{
Converts a data.table in PEcAn CF format to BioCro input format
}
\author{
Rob Kooper, David LeBauer
}

