\name{CheckLeapYear}
\alias{CheckLeapYear}
\title{This fuction checks if year is a leap year. If yes, then returns 1 or 0.}
\usage{
CheckLeapYear(year)
}
\arguments{
  \item{year}{the year in question}
}
\description{
This fuction checks if year is a leap year. If yes, then
returns 1 or 0.
}
\examples{
year==2000 => 1
year==1998 => 0
}

