\docType{data}
\name{cmi0506}
\alias{cmi0506}
\title{Weather data}
\format{Data frame of dimensions 730 by 11.}
\source{
    \url{http://www.sws.uiuc.edu/warm/}
}
\description{
    Selected weather data corresponding to the Champaign weather station (IL,
    U.S.). It has two years: 2005 and 2006. Dimensions: 730 by 11. The columns
    correspond to the input necessary for the \code{\link{weach}} function.
}
\keyword{datasets}
