\name{cmi_weather_data}

\docType{data}

\alias{weather1995}
\alias{weather1996}
\alias{weather1997}
\alias{weather1998}
\alias{weather1999}
\alias{weather2000}
\alias{weather2001}
\alias{weather2002}
\alias{weather2003}
\alias{weather2004}
\alias{weather2005}
\alias{weather2006}
\alias{weather2007}
\alias{weather2008}
\alias{weather2009}
\alias{weather2010}
\alias{weather2011}
\alias{weather2012}
\alias{weather2013}
\alias{weather2014}
\alias{weather2015}
\alias{weather2016}
\alias{weather2017}
\alias{weather2018}
\alias{weather2019}
\alias{weather2020}

\title{Champaign, IL weather data}

\description{
  Multiple years of Champaign, IL weather data specified at hourly intervals in
  the CDT time zone. The columns correspond to typical inputs required for
  BioCro simulations. Note: some values are missing near the start of 1995 since
  the data is not available from SURFRAD.
}

\format{
  26 data frames corresponding to 26 years of weather data, where each data
  frame has 8 columns and 8760 rows.
}

\source{
  Weather data were obtained from the public SURFRAD and WARM databases and
  processed according to the method described in the BioCro II paper (in
  preparation).
}

\keyword{weather}
