% Generated by roxygen2 (4.0.2): do not edit by hand
\name{document}
\alias{document}
\title{devtools document function when using doxygen}
\usage{
document()
}
\value{
The value returned by devtools::document()
}
\description{
Overwrites devtools::document() to include the treatment of
   doxygen documentation in src/
}

