\name{doy124}
\docType{data}
\alias{doy124}
\title{Weather data}
\description{
Example for a given day of the year to illustrate the \code{\link{CanA}} function.

LAI: leaf area index.

year: year.

doy: 124 in this case.

hour: hour of the day, (0--23).

solarR: direct solar radiation.

DailyTemp.C: hourly air temperature (Celsius).

RH: relative humidity, (0--1).

WindSpeed: 4.1 m \eqn{s^{-1}} average daily value in this case.
}

\usage{doy124}
\format{data frame of dimensions 24 by 8.}
\source{simulated}
\keyword{datasets}