% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{data}
\name{doy124}
\alias{doy124}
\title{Weather data}
\format{data frame of dimensions 24 by 8.}
\source{
simulated
}
\description{
Example for a given day of the year to illustrate the \code{\link{CanA}}
function.
}
\details{
LAI: leaf area index.

year: year.

doy: 124 in this case.

hour: hour of the day, (0--23).

solarR: direct solar radiation.

DailyTemp.C: hourly air temperature (Celsius).

RH: relative humidity, (0--1).

WindSpeed: 4.1 m \eqn{s^{-1}} average daily value in this case.
}
\keyword{datasets}

