\docType{data}
\name{doy124}
\alias{doy124}
\title{Weather data}
\format{Data frame of dimensions 24 by 8.}
\source{
    Simulated.
}
\description{
    Example for a given day of the year to illustrate the \code{\link{CanA}}
    function.
}
\details{
    LAI: Leaf area index.

    year: Year.

    doy: 124 in this case.

    hour: Hour of the day ([0, 23]).

    solarR: Direct solar radiation.

    DailyTemp.C: Hourly air temperature (degrees Celsius).

    RH: Relative humidity ([0, 1]).

    WindSpeed: 4.1 m s^-1. mean daily value in this case.
}
\keyword{datasets}
