\name{flow}
\alias{flow}
\title{Returns values based on if kno is less than equal to or greater than three.}
\usage{
flow(SC, CNratio, A, Lc, Tm, resp, kno, Ks, verbose = FALSE)
}
\arguments{
  \item{SC}{}

  \item{CNratio}{}

  \item{A}{}

  \item{Lc}{}

  \item{TM}{}

  \item{resp}{}

  \item{kno}{}

  \item{Ks}{}

  \item{verbose}{Only used in the R version for debugging}
}
\description{
Returns Values for SC, fC, Resp, Kf, and MinN to be used in
the Century function
}

