% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MaizeGro.R
\name{MaizeSeneParms}
\alias{MaizeSeneParms}
\title{Lists values for senecence parameters}
\usage{
MaizeSeneParms(senStem = 3000, senLeaf = 3500, senRoot = 4000)
}
\arguments{
\item{senStem}{senecence of stem}

\item{senLeaf}{senecence of leaf}

\item{senRoot}{senecence of root}
}
\description{
Lists values for senecence parameters
}
