% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/met2model.BIOCRO.R
\name{met2model.BIOCRO}
\alias{met2model.BIOCRO}
\title{Write BioCro met files}
\usage{
met2model.BIOCRO(in.path, in.prefix, outfolder, overwrite = FALSE, ...)
}
\arguments{
\item{in.path}{path on disk where CF file lives}

\item{in.prefix}{prefix for each file}

\item{outfolder}{location where model specific output is written.}

\item{...}{can pass lat, lon, start.date and end.date}
}
\value{
OK if everything was succesful.
}
\description{
Converts a met CF file to a model specific met file. The input
files are calld <in.path>/<in.prefix>.YYYY.cf
}
\author{
Rob Kooper, David LeBauer
}

