% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model2netcdf.BIOCRO.R
\name{model2netcdf.BIOCRO}
\alias{model2netcdf.BIOCRO}
\title{Convert BioCro output to standard netCDF format}
\usage{
model2netcdf.BIOCRO(result, genus = NULL, outdir, lat = -9999, lon = -9999)
}
\arguments{
\item{result}{a dataframe of model output to be converted}

\item{genus}{character: What kind of plant was being simulated?
Used to correct for some genus-specific differences in model output;
Eventually that will be handled inside BioCro and this argument will be removed.}

\item{outdir}{Location of model output}

\item{lat}{Latitude of the site}

\item{lon}{Longitude of the site}
}
\description{
Modified from on model2netcdf.SIPNET and model2netcdf.ED2 by
}
\author{
David LeBauer, Deepak Jaiswal, Rob Kooper
}
