% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/model2netcdf.BIOCRO.R
\name{model2netcdf.BIOCRO}
\alias{model2netcdf.BIOCRO}
\title{Function to convert biocro model output to standard netCDF format}
\usage{
model2netcdf.BIOCRO(result, genus = NULL, outdir, lat = -9999,
  lon = -9999)
}
\arguments{
\item{outdir}{Location of model output}

\item{lat}{Latitude of the site}

\item{lon}{Longitude of the site}
}
\description{
Convert BioCro output to netCDF
}
\details{
Converts BioCro output to netCDF.
Modified from on model2netcdf.SIPNET and model2netcdf.ED2 by
}
\author{
David LeBauer, Deepak Jaiswal, Rob Kooper
}

