% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biocro-package.R
\docType{data}
\name{obsBea}
\alias{obsBea}
\title{Miscanthus assimilation field data}
\format{Data frame of dimensions 27 by 4.}
\source{
C. V. Beale, D. A. Bint, S. P. Long. 1996. Leaf photosynthesis in the
C4-grass Miscanthus x giganteus, growing in the cool temperate climate of
southern England. \emph{J. Exp. Bot.} 47 (2): 267--273.
}
\description{
Assimilation in Miscanthus as measured in Beale, Bint, and Long 1996.  The
first column is the observed net assimilation rate (\eqn{\mu} mol
m\eqn{^{-2}}{-2} s\eqn{^{-1}}{-1}).  The second column is the observed quantum flux
(\eqn{\mu} mol m\eqn{^{-2}}{-2} s\eqn{^{-1}}{-1}). The third column is the
temperature (Celsius).  Relative humidity was not reported and thus was
assumed to be 0.7.
}
\keyword{datasets}
