\docType{data}
\name{obsBea}
\alias{obsBea}
\title{Miscanthus assimilation field data}
\format{Data frame of dimensions 27 by 4.}
\source{
    C. V. Beale, D. A. Bint, S. P. Long. 1996. Leaf photosynthesis in the
    C4-grass Miscanthus x giganteus, growing in the cool temperate climate of
    southern England. \emph{J. Exp. Bot.} 47 (2): 267--273.
}
\description{
    Assimilation in Miscanthus as measured in Beale, Bint, and Long 1996.  The
    first column is the observed net assimilation rate (micromoles m^-2 s^-1).
    The second column is the observed quantum flux (micromoles m^-2 s^-1).
    The third column is the temperature (degrees Celsius). Relative humidity
    was not reported and thus was assumed to be 0.7.
}
\keyword{datasets}
