\name{obsBeaC}
\docType{data}
\alias{obsBeaC}
\title{Complete Miscanthus assimilation field data}
\description{
  
Assimilation and stomatal conductance as measured in Beale, Bint and
Long (1996) in Miscanthus (missing data are also included).  The first column is the date, the second the
hour. Columns 3 and 4 are assimilation and stomatal conductance
respectively.
  
 The third column is the observed net assimilation rate (\eqn{\mu} mol
\eqn{m^{-2}} \eqn{s^{-1}}).

The fifth column is the observed quantum flux  (\eqn{\mu} mol \eqn{m^{-2}}
\eqn{s^{-1}}).

The sixth  column is the temperature (Celsius).
}
\usage{obsBeaC}
\format{data frame of dimensions 35 by 6.}
\source{
C. V. Beale, D. A. Bint, S. P. Long, Leaf photosynthesis in the C4-grass miscanthus
x giganteus, growing in the cool temperate climate of southern England,
\emph{J. Exp. Bot.} 47 (2) (1996) 267--273.
}
\keyword{datasets}