% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biocro-package.R
\docType{data}
\name{obsNaid}
\alias{obsNaid}
\title{Miscanthus assimilation data}
\format{data frame of dimensions 16 by 4.}
\source{
S. L. Naidu, S. P. Moose, A. K. AL-Shoaibi, C. A. Raines, S. P.
Long, Cold Tolerance of C4 photosynthesis in Miscanthus x giganteus:
Adaptation in Amounts and Sequence of C4 Photosynthetic Enzymes,
\emph{Plant Physiol.} 132 (3) (2003) 1688--1697.
}
\description{
assimilation as measured in Naidu et al. (2003) in Miscanthus. The first
column is the observed net assimilation rate (\eqn{\mu} mol \eqn{m^{-2}}
\eqn{s^{-1}}) The second column is the observed quantum flux (\eqn{\mu} mol
\eqn{m^{-2}} \eqn{s^{-1}}) The third column is the temperature (Celsius).
The fourth column is the observed relative humidity in proportion (e.g.
0.7).
}
\keyword{datasets}
