\docType{data}
\name{obsNaid}
\alias{obsNaid}
\title{Miscanthus assimilation data}
\format{Data frame of dimensions 16 by 4.}
\source{
    S. L. Naidu, S. P. Moose, A. K. AL-Shoaibi, C. A. Raines, S. P.
    Long.  2003. Cold Tolerance of C4 photosynthesis in Miscanthus x giganteus:
    Adaptation in Amounts and Sequence of C4 Photosynthetic Enzymes.
    \emph{Plant Physiol.} 132 (3): 1688--1697.
}
\description{
    Assimilation in Miscanthus as measured in Naidu et al. (2003). The first
    column is the observed net assimilation rate (micromoles m^-2 s^-1).
    The second column is the observed quantum flux (micromoles m^-2 s^-1).
    The third column is the temperature (degrees Celsius).
    The fourth column is the observed relative humidity in proportion (e.g. 0.7).
}
\keyword{datasets}
