% Generated by roxygen2 (4.0.1): do not edit by hand
\name{read.config.BIOCRO}
\alias{read.config.BIOCRO}
\title{Read BioCro Config}
\usage{
read.config.BIOCRO(config.file = "config.xml")
}
\arguments{
\item{config.file}{}
}
\value{
list of run configuration parameters for PEcAn
}
\description{
Read BioCro config file
}
\details{
TODO: Integrate into BioCro itself
}
\author{
David LeBauer
}

