\name{rootDist}
\alias{rootDist}
\title{Returns a value for rootDist based on the arguments layers, rootDepth, depthsp, and rfl.
This value then factors into several equations in the primary function soilML.}
\usage{
rootDist(layers, rootDepth, depthsp, rfl)
}
\description{
Returns a value for rootDist based on the arguments layers,
rootDepth, depthsp, and rfl. This value then factors into
several equations in the primary function soilML.
}

