% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RssBioGro.R
\name{RssBioGro}
\alias{RssBioGro}
\title{Residual sum of squares for BioGro.}
\usage{
RssBioGro(obs, sim)
}
\arguments{
\item{obs}{Observed data.}

\item{sim}{Simulated data.}
}
\value{
Atomic vector with the residual sum of squares.
}
\description{
Computes residual sum of squares for the \code{\link{BioGro}} function.
}
\examples{

## A simple example
data(annualDB)
data(EngWea94i)
res <- BioGro(EngWea94i)
RssBioGro(annualDB,res)

}
\seealso{
See Also \code{\link{BioGro}}.
}
\author{
Fernando E. Miguez
}
\keyword{models}
