% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/run.biocro.R
\name{run.biocro}
\alias{run.biocro}
\title{Run BioCro at a point}
\usage{
run.biocro(lat, lon, met.nc = met.nc, soil.nc = NULL, config = config,
  coppice.interval = 1)
}
\arguments{
\item{lat}{latitude in decimal degrees}

\item{lon}{longitude in decimal degrees}

\item{met.nc}{full path and name of a netCDF file in PEcAn-CF format with meteorological driver data}

\item{soil.nc}{full path and name of a netCDF file with soil data}

\item{config}{full path and name of a config.xml file containing parameter values and configuration information for BioCro}

\item{coppice.interval}{numeric, number of years between cuttings for coppice plant or perinneal grass (default 1)}
}
\value{
output from one of the \code{BioCro::*.Gro} functions (determined by \code{config$genus}), as data.table object
}
\description{
Run BioCro at a point
}
\author{
David LeBauer
}

