\name{soybean2}

\docType{data}

\alias{soybean2}

\title{Soybean-BioCro model definition}

\description{
  Alternate version of the \code{\link{soybean}} model, where the main
  difference is that \code{soybean2} uses a new set of soil water modules.
}

\usage{soybean2}

\format{
  A list of 5 named elements that are suitable for passing to
  \code{\link{run_biocro}}, as described in the help page for
  \code{\link{crop_model_definitions}}.
}

\seealso{
  \itemize{
    \item \code{\link{run_biocro}}
    \item \code{\link{modules}}
    \item \code{\link{crop_model_definitions}}
    \item \code{\link{soybean_clock}}
  }
}

\keyword{datasets}

\concept{crop_models}
