% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/start.runs.BIOCRO.R
\name{start.runs.BIOCRO}
\alias{start.runs.BIOCRO}
\title{Start run of biocro model}
\usage{
\method{start}{runs.BIOCRO}(runid)
}
\arguments{
\item{runid}{the id of the run (folder in runs) to execute}
}
\description{
Start biocro model runs on local
}
\author{
Rob Kooper, David LeBauer, Deepak Jaiswal
}

