% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Opc4photo.R
\name{summary.Opc4photo}
\alias{summary.Opc4photo}
\title{This function will implement simple calculations
of predicted and residuals for the Opc4photo function}
\usage{
\method{summary}{Opc4photo}(object, ...)
}
\description{
This function will implement simple calculations
of predicted and residuals for the Opc4photo function
}
