\name{TempToDdryA}
\alias{TempToDdryA}
\title{takes a value for Temp as defined by the SoilEvapo function and returns a value for DdryA
which factors into variable PsychParam which in turn helps define the Evaporation.}
\usage{
TempToDdryA(Temp)
}
\description{
takes a value for Temp as defined by the SoilEvapo function
and returns a value for DdryA which factors into variable
PsychParam which in turn helps define the Evaporation.
}

