\name{TempToLHV}
\alias{TempToLHV}
\title{Returns a value for TempToLHV}
\usage{
TempToLHV(Temp)
}
\arguments{
  \item{Temp}{Temperature}
}
\description{
Takes a value for Temp as defined by the SoilEvapo function
and returns a value for LHV which factors into variable
PsychParam which in turn helps define the Evaporation.
}

