% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/SoilEvapo.R
\name{TempToSFS}
\alias{TempToSFS}
\title{Returns a value for SlopeFS from a function of temperature}
\usage{
TempToSFS(Temp)
}
\arguments{
\item{Temp}{Temperature}
}
\description{
Takes a value for Temp as defined by the SoilEvapo function and returns a value for SlopeFS
which helps define the Evaporation.
}

