\name{TempToSFS}
\alias{TempToSFS}
\title{Returns a value for SlopeFS as a funtion of temperature}
\usage{
TempToSFS(Temp)
}
\arguments{
  \item{Temp}{Temperature}
}
\description{
Takes a value for Temp as defined by the SoilEvapo function
and returns a value for SlopeFS which helps define the
Evaporation.
}

