\name{TempToSWVC}
\alias{TempToSWVC}
\title{takes a value for Temp as defined by the SoilEvapo function and returns a value for SWVC
which factors into variable DeltaPVa which in turn helps define the Evaporation.}
\usage{
TempToSWVC(Temp)
}
\description{
takes a value for Temp as defined by the SoilEvapo function
and returns a value for SWVC which factors into variable
DeltaPVa which in turn helps define the Evaporation.
}

