\docType{data}
\name{weather05}
\alias{weather05}
\alias{weather04}
\title{Weather data}
\format{Data frame of dimensions 8760 by 7.}
\source{
    Simulated data based on Champaign, Illinois conditions.
}
\description{
    Weather data as produced by the \code{\link{weach}} function.  These are
    for 2004 and 2005.
}
\keyword{datasets}
