% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biocro-package.R
\docType{data}
\name{weather06}
\alias{weather06}
\title{Weather data}
\format{A data frame with 8760 observations on the following 8 variables.
\describe{ \item{list('year')}{a numeric vector} \item{list('doy')}{a
numeric vector} \item{list('hour')}{a numeric vector}
\item{list('SolarR')}{a numeric vector} \item{list('Temp')}{a numeric
vector} \item{list('RH')}{a numeric vector} \item{list('WS')}{a numeric
vector} \item{list('precip')}{a numeric vector} }}
\source{
%% ~~ reference to a publication or URL from which the data were
obtained ~~
}
\description{
Weather data as produced by the \code{\link{weach}} function.  These are
for 2006.
}
\examples{

data(weather06)
## maybe str(weather06) ; plot(weather06) ...

}
\keyword{datasets}
